Move NO_DATA SYMBOL to common + rename module
