+++ frontend/src/components/dashboard/table-common.tsx
@@ -1,3 +1,5 @@
+export const NO_DATA = "--" as const
+
 export enum ActiveDialog {
   FolderRename,
   FolderDelete,
